Moved the app theme into its own module (theme.ts) and added default styling for Center components (which wrap all my screens)
