--- src/ui/helpers.ts
@@ -1,6 +1,0 @@
-// Helpers for styling UI components
-import { EdgeInsets } from "react-native-safe-area-context";
-
-export const getSafePadding = (insets: EdgeInsets): string => {
-  return `${insets.top}px ${insets.right}px ${insets.bottom}px ${insets.left}px`;
-};
