+++ src/ui/index.ts
@@ -1,3 +1,4 @@
 export { Home } from "./Home/Home";
 export { RecipeForm } from "./RecipeForm/RecipeForm";
 export { RecipeView } from "./RecipeView/RecipeView";
+export { THEME } from "./theme";
